+++ crates/cargo-zng/src/res.rs
@@ -178,10 +178,7 @@ fn source_to_target_pass(args: &ResArgs, tools: &Tools, source: &Path, target: &
                 let ext = ext.to_string_lossy();
                 if let Some(tool) = ext.strip_prefix("zr-") {
                     // run prints request
-                    let output = tools.run(tool, &args.source, &args.target, source)?;
-                    for line in output.lines() {
-                        println!("  {line}");
-                    }
+                    tools.run(tool, &args.source, &args.target, source)?;
                     continue;
                 }
             }
@@ -214,10 +211,7 @@ fn target_to_target_pass(args: &ResArgs, tools: &Tools, dir: &Path) -> anyhow::R
                     // run prints request
                     let tool_r = tools.run(tool, &args.source, &args.target, path);
                     fs::remove_file(path)?;
-                    let output = tool_r?;
-                    for line in output.lines() {
-                        println!("  {line}");
-                    }
+                    tool_r?;
                 }
             }
         }
@@ -237,7 +231,6 @@ fn tools_help(tools: &Path) {
             }
             Err(e) => error!("{e}"),
         }
-
         Ok(ControlFlow::Continue(()))
     });
     if let Err(e) = r {